netrw: avoid clobbering filetype when opening netrw
Some filetypes are set manually, or using autocmd, and do not work
properly with `filetype detect`, so in netrw, save and restore the
buffer's current filetype to be used in place of `filetype detect` in
`NetrwOptionRestore()`.

The original `filetype detect` call is left in place to account for the
possibility that `{a:vt}netrw_ftkeep` is not set.